fix wrong data type in todayinfo model

+++ src/models.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 
 class TodayInfo(BaseModel):
     puzzleNumber: int
-    similarity: int
+    similarity: float
     similarityTenth: float
     similarityThousandth: float
